requirements.txt
fastapi
uvicorn[standard]
numpy>=2.1.0
scipy==1.13.1
pillow==10.4.0
python-dotenv==1.0.1
pandas==2.2.0
Jinja2 